+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## UPCOMING
+
+- Feature: `CLIENT KILL` now support `LADDR` argument.
+-
 ## 1.11.1
 
 - Bugfix: Switch `distutils.version` to `packaging.version` to fix the version parse